Renames image_util package to swri_image_util. Refs #231.
