--- image_util/launch/geometry_util.test
@@ -1,3 +1,0 @@
-<launch>
-  <test test-name="test_geometry_util" pkg="image_util" type="test_geometry_util" />
-</launch>
