--- image_util/nodelet_plugins.xml
@@ -1,27 +1,0 @@
-<library path="lib/libimage_util_nodelets">
-  
-  <class name="image_util/rotate_image" type="image_util::RotateImageNodelet" base_class_type="nodelet::Nodelet">
-    <description>
-      Nodelet to rotate an image a multiple of 90 degrees.
-    </description>
-  </class>
-  
-  <class name="image_util/scale_image" type="image_util::ScaleImageNodelet" base_class_type="nodelet::Nodelet">
-    <description>
-      Nodelet to scale an image.
-    </description>
-  </class>
-  
-  <class name="image_util/contrast_stretch" type="image_util::ContrastStretchNodelet" base_class_type="nodelet::Nodelet">
-    <description>
-      Nodelet to contrast stretch an image.
-    </description>
-  </class>
-  
-  <class name="image_util/draw_text" type="image_util::DrawTextNodelet" base_class_type="nodelet::Nodelet">
-    <description>
-      Nodelet to draw text on an image.
-    </description>
-  </class>
-  
-</library>
