--- image_util/src/rolling_normalization.cpp
@@ -1,99 +1,0 @@
-// *****************************************************************************
-//
-// Copyright (c) 2014, Southwest Research Institute® (SwRI®)
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
-//     * Redistributions of source code must retain the above copyright
-//       notice, this list of conditions and the following disclaimer.
-//     * Redistributions in binary form must reproduce the above copyright
-//       notice, this list of conditions and the following disclaimer in the
-//       documentation and/or other materials provided with the distribution.
-//     * Neither the name of Southwest Research Institute® (SwRI®) nor the
-//       names of its contributors may be used to endorse or promote products
-//       derived from this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-// ARE DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> BE LIABLE FOR ANY
-// DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-// (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-// LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-// ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-// *****************************************************************************
-
-#include <image_util/rolling_normalization.h>
-
-namespace image_util
-{
-  RollingNormalization::RollingNormalization(int32_t size) :
-    max_size_(size),
-    samples_(0)
-  {
-    
-  }
-  
-  RollingNormalization::~RollingNormalization()
-  {
-  
-  }
-    
-  cv::Mat RollingNormalization::AddSample(const cv::Mat& image)
-  {
-    if (samples_ == 0)
-    {
-      image.convertTo(average_image_, CV_64F, 1.0, 0.0);
-    }
-    else
-    {
-      cv::Mat temp;
-      image.convertTo(temp, CV_64F, 1.0, 0.0);
-      double s = static_cast<double>(samples_);
-      average_image_ = (average_image_ * s + temp) / (s + 1.0);
-    }
-    
-    samples_++;
-    if (samples_ > max_size_)
-    {
-      samples_ = max_size_;
-    }
-    
-    cv::Mat mean_image;
-    average_image_.convertTo(mean_image, CV_8U);
-    cv::Mat temp_norm_image;
-
-    cv::medianBlur(mean_image, temp_norm_image, 25);
-
-    cv::Mat temp_norm_image2;
-    temp_norm_image.convertTo(temp_norm_image2, CV_32F);
-    double max1 = 0;
-    for (int32_t i = 0; i < temp_norm_image2.rows; i++)
-    {
-      for (int32_t j = 0; j < temp_norm_image2.cols; j++)
-      {
-        if (temp_norm_image2.at<float>(i, j) > max1)
-        {
-          max1 = temp_norm_image2.at<float>(i, j);
-        }
-      }
-    }
-
-    temp_norm_image2 = temp_norm_image2 * (255.0 / max1);
-
-    cv::Mat temp_norm_image3;
-    temp_norm_image2.convertTo(temp_norm_image3, CV_8U, 1.0, 0.0);
-
-
-    cv::GaussianBlur(temp_norm_image3,
-                     norm_image_,
-                     cv::Size(15, 15),
-                     5,
-                     5);
-    return norm_image_;
-  }
-}
