+++ swri_image_util/mainpage.dox
