+++ swri_image_util/package.xml
@@ -1,0 +1,40 @@
+<package format="2">
+  <name>swri_image_util</name>
+  <version>0.0.1</version>
+  <description>
+
+     swri_image_util
+
+  </description>
+  <author>Kris Kozak</author>
+  <maintainer email="kkozak@swri.org">Kris Kozak</maintainer>
+  <license>BSD</license>
+  <url>https://github.com/swri-robotics/marti_common</url>
+ 
+  <buildtool_depend>catkin</buildtool_depend>
+
+  <depend>camera_calibration_parsers</depend>
+  <depend>cmake_modules</depend>
+  <depend>cv_bridge</depend>
+  <depend>eigen</depend>
+  <depend>geometry_msgs</depend>
+  <depend>image_geometry</depend>
+  <depend>image_transport</depend>
+  <depend>libqt4-dev</depend>
+  <depend>math_util</depend>
+  <depend>nav_msgs</depend>
+  <depend>nodelet</depend>
+  <depend>opencv_util</depend>
+  <depend>roscpp</depend>
+  <depend>roslib</depend>
+  <depend>rospy</depend>
+  <depend>std_msgs</depend>
+  <depend>tf</depend>
+
+  <test_depend>rostest</test_depend>
+
+  <export>
+    <nodelet plugin="${prefix}/nodelet_plugins.xml" />
+  </export>
+</package>
+
