+++ swri_image_util/src/nodes/dummy_image_publisher.cpp
